Minor updates
1. When querying all nodes from a checklist, the
previous recursive method would create and destroy a
number of std::vector objects.  Now we're querying
iteratively and no longer creating extra objects
2. Fixed the means by which the icon bundles are
created.  Separate methods are needed between
windows and linux due to the whole resource thing
needed to pack the icons directly into the exe on
windows
3. Moved some constants being used (default border
and flags used for adding sizers to the main sizer) so
it's easier to change them
4. Since windows and linux need entirely separate
signatures for creating the icon bundle, that's been
moved into the main window.cpp